+++ BTCPayServer/Plugins/Bitpay/Security/TokenRepository.cs
@@ -36,52 +36,6 @@ namespace BTCPayServer.Plugins.Bitpay.Security
                 .ToArray();
         }
 
-        public async Task<String> GetStoreIdFromAPIKey(string apiKey)
-        {
-            using var ctx = _Factory.CreateContext();
-            return await ctx.ApiKeys.Where(o => o.Id == apiKey).Select(o => o.StoreId).FirstOrDefaultAsync();
-        }
-
-        public async Task GenerateLegacyAPIKey(string storeId)
-        {
-            // It is legacy support and Bitpay generate string of unknown format, trying to replicate them
-            // as good as possible. The string below got generated for me.
-            var chars = "ERo0vkBMOYhyU0ZHvirCplbLDIGWPdi1ok77VnW7QdE";
-            var generated = new char[chars.Length];
-            for (int i = 0; i < generated.Length; i++)
-            {
-                generated[i] = chars[(int)(RandomUtils.GetUInt32() % generated.Length)];
-            }
-
-            using var ctx = _Factory.CreateContext();
-            var existing = await ctx.ApiKeys.Where(o => o.StoreId == storeId && o.Type == APIKeyType.Legacy).ToListAsync();
-            if (existing.Any())
-            {
-                ctx.ApiKeys.RemoveRange(existing);
-            }
-            ctx.ApiKeys.Add(new APIKeyData() { Id = new string(generated), StoreId = storeId });
-            await ctx.SaveChangesAsync().ConfigureAwait(false);
-        }
-
-        public async Task RevokeLegacyAPIKeys(string storeId)
-        {
-            var keys = await GetLegacyAPIKeys(storeId);
-            if (!keys.Any())
-            {
-                return;
-            }
-
-            using var ctx = _Factory.CreateContext();
-            ctx.ApiKeys.RemoveRange(keys.Select(s => new APIKeyData() { Id = s }));
-            await ctx.SaveChangesAsync();
-        }
-
-        public async Task<string[]> GetLegacyAPIKeys(string storeId)
-        {
-            using var ctx = _Factory.CreateContext();
-            return await ctx.ApiKeys.Where(o => o.StoreId == storeId && o.Type == APIKeyType.Legacy).Select(c => c.Id).ToArrayAsync();
-        }
-
         private BitTokenEntity CreateTokenEntity(PairedSINData data)
         {
             return new BitTokenEntity()